Refactored and included documentation
