+++ .gitignore
@@ -7,6 +7,7 @@
 
 # testing
 /coverage
+/src/fixtures
 
 # production
 /build
