+++ src/App.css
@@ -4,7 +4,6 @@ footer
     background-color: #363537;
     color: white;
     align-items: center;
-    /*text-align: center;*/
 }
 footer > div > h1, p
 {
