+++ src/components/Menu/Category/MenuItem/MenuItem.css
@@ -1,0 +1,30 @@
+.menuitem-container
+{
+    border: 0px solid black;
+    background-color: white;
+    width: 100%;
+    height: 5em;
+    overflow: hidden;
+    padding: 0.5em;
+    display: flex;
+}
+
+.menuitem-details
+{
+    margin-left: 1em;
+    width: 80%;
+}
+
+.menuitem-img-container
+{
+    border: 2px solid black;
+    width: 4em;
+    height: 4em;
+}
+
+.menuitem-label > h3
+{
+    display: inline-block;
+    text-align: center;
+    text-transform: capitalize;
+}
