+++ src/utilities.css
@@ -1,3 +1,7 @@
+/**
+ * This file contains classes that are used across multiple components.
+ */
+
 @import url("https://fonts.googleapis.com/css2?family=Dancing+Script:wght@562&family=Lobster&family=Lobster+Two&family=Norican&family=Pacifico&display=swap");
 
 * {
@@ -25,7 +29,6 @@
         
     }
     
-
 .imgContainer-circle
 {
     border-radius: 50%;
